refactor(agent): extract stateless free helpers from dkg-agent.ts
Moves 21 pure free functions out of `dkg-agent.ts` into a new
`dkg-agent-helpers.ts` module, grouped by purpose:

  - Publish-payload normalisation (`normalizePublishContextGraphId`,
    `isPublishAsyncQuadEnvelope`, `assertQuadArray`,
    `partitionPublishAsyncQuads`).
  - EIP-712 sign + attestation byte ↔ hex conversions
    (`signWithPrivateKey`, `preSignedAttestationToLiftSeal`).
  - DID / delegation-scope normalisation (`normalizeAgentDid`,
    `joinDelegationScope`).
  - Sync-phase normalisation (`normalizeSyncPhase`).
  - ChainAdapter publisher-address discovery
    (`normalizeAdapterPublisherAddress`, `recoverCompactSigner`,
    `adapter{OperationalPrivateKeyAddress,HasOperationalPrivateKey,
    GenericSignMessageMatchesAddress,AdvertisesPublisherSigner}`,
    `privateKeyAddress`, `inferAdapterPublisherAddress`).
  - Triple-store config defaults (`defaultLargeLiteralStorage`,
    `createPublicSnapshotStore`, `applyDefaultLargeLiteralStorage`,
    `isLocalOxigraphConfig`).

None of these touch `DKGAgent` state — every one was already a free
function reading its inputs explicitly. The move is purely mechanical;
JSDoc and behaviour are preserved.
